MEDIUM: server: Allow generic type-value pairs as proxy-v2-options
This commit removes the previous limitations on the existing, fixed PPv2 TLV types
that can be sent out. It introduces a generic server-side parsing of type-value pair
arguments and allocation of a TLV list. This allows to 1) forward any TLV type,
2) generally send out any TLV type, and 3) allows to specify concrete values via
an '=' after the type and a log fmt expression. If there is no TLV found in the
remote connection, an empty TLV is sent out.