MEDIUM: tcp-act: Add new set-tlv TCP action for PPv2 TLVs
This commit adds an action called set-tlv(<id>) <expr> that allows to directly
update the TLV data structure within a connection for all type of connection
events. It can be used to modify TLVs before they are forwarded (if specified
in proxy-v2-options), keeping the previously allocated memory. Besides, if a
TLV does not exist, it is created and added to the list.