MEDIUM: tcp-act: Add new set-tlv TCP action for PPv2 TLVs
This commit adds an action called set-tlv(<id>) <expr> that allows to directly
update the TLV data structure within a connection for all type of connection
events. It can be used to modify TLVs before they are forwarded (if specified
in proxy-v2-options) while keeping the previously allocated memory, if the new
and the old value map to the same pool. This function can also be used to
enhance readability if setting many TLVs at once, as an alternative to specifying
type and value directly in the server.