MEDIUM: server: Send out generically allocated proxy-v2-options
This commit removes the previous limitations on the existing, fixed PPv2 TLV types
that can be sent out. This is done by leveraging the previously implemented
parsing. We iterate over the server TLV list and either forward or send a new
TLV, depending on whether a TLV exists in the remote connection.