BUG/MINOR: mworker/cli: don’t skip same‑timestamp workers for "show proc"
After commit 594408c, related to ticket #3204, the "show proc" logic has
been fixed to be able to print more than 202 processes. However, this
fix can lead to the omission of entries in case they have the same
timestamp.

To fix this, we now flush the CLI output only when the worker timestamp
changes, batching entries that share the same second. We apply the same
logic to old workers and update ctx->next_uptime at each flush and after
the final group to resume correctly.